Fix `dbconn` dependency version for backward-compatibility
`dbconn` version `v1.0.3` broke compatibility by removing the global `DB` object to force the use of the `connect_to_db()`  factory function. As the `dbconn` dependency in `utils` was not versioned, new builds triggered after the new `dbconn` version release automatically got the new version and broke because of the lack of the `DB` object.

To allow gradual functions migration to the new `dbconn` API, create a parallel `utils` version branch with fixed `dbconn` dependency with the latest version before `DB` removal. Functions not yet migrated may target this parallel `utils` versions to keep using a compatible `dbconn` version.

In the future, it would be better to remove the `dbconn` dependency from `utils`, moving everything that depends on it (e.g. `session_factory()`) to the `dbconn` package itself.